requirements.txt
apispec ~= 0.27.0
flask ~= 0.12.2
flask-cors ~= 3.0.3
jinja2 ~= 2.10
jsonpatch ~= 1.21
jsonschema ~= 2.6
pycontracts ~= 1.8.0
